Thêm ảnh Adpat giao diện

+++ FoodCourtSystem/Models/CartViewModel.cs
@@ -9,9 +9,9 @@ namespace FoodCourtSystem.Models
     public class CartItemModel
     {
         public int Quantity { get; set; }
-        public double TotalMoney { get; set;}
+        public int TotalMoney { get; set;}
         public ProductModel Product { get; set; }
-        private double getTotalMoney()
+        private int getTotalMoney()
         {
             return Product.UnitPrice * Quantity;
         }
@@ -22,7 +22,7 @@ namespace FoodCourtSystem.Models
         public string OwnerName { get; set; }
         public virtual ICollection<CartItemModel> items { get; set; }
         public double VAT { get;}
-        public double TotalMoney { get; set; }
+        public int TotalMoney { get; set; }
         public CartModel()
         {
             VAT = 0.1;
@@ -33,7 +33,7 @@ namespace FoodCourtSystem.Models
             double price = 0;
             foreach (CartItemModel item in items)
                 price += item.TotalMoney;
-            TotalMoney = price * (1 + VAT);
+            TotalMoney = (int)((double)price * (1 + VAT));
         }
     }
